Removed lithium.cabal file, hpack generates this
Developers with more recent versions of either hpack or stack are
receiving warning messages during build:

> Warning: /home/smedley/lithium/lithium.cabal was generated with a
> newer version of hpack, please upgrade and try again.

This is because, even though this project uses hpack by having a
`package.yaml` file, it also has a checked-in `lithium.cabal` file. I
believe the pattern of use is to either manage the cabal file yourself
or let hpack do it, but not have both a `.cabal` and a `package.yaml`
file in the repo.

This commit removes the `lithium.cabal` file from the repo, opting to
manage it via hpack. A `.gitignore` rule was also added to ignore cabal
files from here on out. Thanks to @dmp1ce for help troubleshooting this
issue.